requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.1
flask-migrate>=4.0.5
flask-cors>=4.0.0
python-dotenv>=1.0.1
mysqlclient>=2.2.4
pymysql>=1.1.0
reportlab>=4.0.8
google-generativeai>=0.5.0   # optional for Gemini API

